feat(client): support local bind for HttpConnector
Temporary version - awaiting Hyper 0.12 for proper version (see issue 1498).

Add `set_local_address` to the `HttpConnector`.
This configures the client to bind the socket to a local address of
the host before it connects to the destination. This is useful on
hosts which have multiple network interfaces, to ensure the request is
issued over a specific interface.